server/requirements.txt
Flask==1.0.2
gunicorn==19.9.0
google-cloud-speech
google-cloud-language
google-cloud-storage==1.13.2
nltk
google-cloud-datastore==1.7.3
